Migrate equipments page
